refactor: move academic sources into package
