+++ src/nature_academic_search/conversion/__init__.py
@@ -1,0 +1,15 @@
+"""Citation conversion helpers."""
+
+from .converters import (
+    convert_from_arxiv,
+    convert_from_crossref,
+    convert_from_medline,
+    get_extension,
+)
+
+__all__ = [
+    "convert_from_arxiv",
+    "convert_from_crossref",
+    "convert_from_medline",
+    "get_extension",
+]
