+++ src/nature_academic_search/sources/__init__.py
@@ -1,0 +1,7 @@
+"""Data source adapters for academic search."""
+
+from .arxiv import ArxivSource
+from .crossref import CrossRefSource
+from .pubmed import PubMedSource
+
+__all__ = ["CrossRefSource", "PubMedSource", "ArxivSource"]
